switched charsets to be vectors, changed regex and lex_state to assume an alphabet of 256

+++ regex/from_literal.h
@@ -1,4 +1,2 @@
-struct regex* regex_from_literal(
-	const unsigned char* string,
-	size_t len
-);
+
+struct regex* regex_from_literal(const unsigned char* string, unsigned len);
